Complete restructure again. Calling SFDP is now an internal function. The library now gathers all sfdp information through _chipID(). SFDP information is collected when available, otherwise the library defaults to the pre v3.2.0 version.
