+++ examples/TestFlash/TestFlash.ino
@@ -91,10 +91,8 @@ String inputString, outputString;
 #define BAUD_RATE 115200
 #endif
 
-SPIFlash flash;                   //Use this constructor if your Flash chip does not support the SFDP standard - autoidentification will no longer work and you will be restricted to the flash memory chips listed in ReadMe.md
+SPIFlash flash;                   
 //SPIFlash flash(SS1, &SPI1);       //Use this constructor if using an SPI bus other than the default SPI. Only works with chips with more than one hardware SPI bus
-//SFDPFlash flash;                    
-//SFDPFlash flash(SS1, &SPI1);       //Use this constructor if using an SPI bus other than the default SPI. Only works with chips with more than one hardware SPI bus
 
 void setup() {
   Serial.begin(BAUD_RATE);
