+++ extras/Changes.log
@@ -8,12 +8,11 @@
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 --> Library name changed to SPIMemory
 
---> Library now has a different structure. Calling the constructor ```SPIFlash flash;``` initialises the library as it was at version 3.1.0 and preserves legacy compatibility (this is guaranteed till v4.0.0 is released) However, calling the new constructor ```SFDPFlash flash;``` initialises the new and upgraded version of the library with all the enhancements listed in the Enhancements section below. This variant of the library is currently in BETA and will merge into and replace the legacy version of the library (with constructor SPIFlash) in version 4.0.0.
-
---> Both SPIFlash & SFDPFlash variants of the library will receive all bug-fixes till v4.0.0. However all feature enhancements will only be applied to the SFDPFlash variant from v3.2.0 onwards.
+--> The library now gathers sfdp information through _chipID(). SFDP information is collected when available, otherwise the library defaults to the pre v3.2.0 version.
 
 New flash memory chips supported
 --> All flash memories compatible with the SFDP standard (as at JESD216B) can be used with the constructor ```SFDPFlash```
+
 --> LE25U40CMC from ON Semiconductor
 
 Bugs squashed
